--- src/index.ts
@@ -1,5 +1,0 @@
-import Container from './common/Container';
-import FlexBox from './common/FlexBox';
-import Grid from './common/Grid';
-
-export { Container, FlexBox, Grid };
