+++ stories/Container/Container.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import { Container } from '../..';
+import { Container } from '../../lib';
 
 const meta = {
   title: 'Container',
